+++ frontend/src/pages/Child/ChildDetails.jsx
@@ -3,14 +3,14 @@ import { useParams, useNavigate } from "react-router-dom";
 import NavBar from "../../components/NavBar.jsx";
 import { getTextColourFromBrightness } from "../../utils/utils.js";
 import { useAppContext } from "../../context/AppContext.jsx";
+import ScheduledTransactionTable from "../../components/Child/ScheduledTransactionTable.jsx";
 
 export default function ChildPage() {
     const { householdId, childId } = useParams();
     const navigate = useNavigate();
     const [child, setChild] = useState(null);
     const [accounts, setAccounts] = useState([]);
-    const [transactionSchedules, setTransactionSchedules] = useState([]);
-    const { apiFetch, logout } = useAppContext();
+    const { apiFetch } = useAppContext();
 
     useEffect(() => {
         apiFetch(`children/${childId}`)
@@ -38,12 +38,6 @@ export default function ChildPage() {
             })
             .catch((error) =>
                 console.error("Error fetching accounts data:", error),
-            );
-        apiFetch(`children/${childId}/scheduled_transactions`)
-            .then((response) => response.json())
-            .then((data) => setTransactionSchedules(data["member"]))
-            .catch((error) =>
-                console.error("Error fetching transaction schedules:", error),
             );
     }, [childId, navigate]);
 
@@ -80,18 +74,6 @@ export default function ChildPage() {
             `/household/${householdId}/child/${childId}/account/${accountId}/transaction/add`,
         );
     };
-
-    const handleCreateScheduledTransaction = () => {
-        navigate(
-            `/household/${householdId}/child/${childId}/scheduled_transaction/add`,
-        );
-    };
-
-    function handleViewScheduledTransaction(scheduleId) {
-        navigate(
-            `/household/${householdId}/child/${childId}/scheduled_transaction/${scheduleId}`,
-        );
-    }
 
     return (
         <div>
@@ -201,81 +183,6 @@ export default function ChildPage() {
                                     </li>
                                 )}
                             </ul>
-                        </div>
-                    </div>
-                </div>
-
-                {/* Transaction Schedules Table */}
-                <div
-                    className="card shadow-sm border-0 mb-4"
-                    style={{ backgroundColor: "#f8f9fa" }}
-                >
-                    <div className="card-body">
-                        <div className="d-flex justify-content-between align-items-center mb-3">
-                            <h3 className="card-title text-primary">
-                                Scheduled transactions
-                            </h3>
-                            <button
-                                onClick={handleCreateScheduledTransaction}
-                                className="btn btn-success"
-                            >
-                                + Add scheduled transaction
-                            </button>
-                        </div>
-
-                        <div className="mt-3">
-                            <table className="table table-striped">
-                                <thead>
-                                    <tr>
-                                        <th>Next Execution</th>
-                                        <th>Description</th>
-                                        <th>Amount</th>
-                                        <th>Comment</th>
-                                        <th>Actions</th>
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    {transactionSchedules.length > 0 ? (
-                                        transactionSchedules.map((schedule) => (
-                                            <tr key={schedule["@id"]}>
-                                                <td>
-                                                    {new Date(
-                                                        schedule.nextExecutionDate,
-                                                    ).toLocaleDateString()}
-                                                </td>
-                                                <td>{schedule.description}</td>
-                                                <td>
-                                                    $
-                                                    {schedule.amount.toFixed(2)}
-                                                </td>
-                                                <td>{schedule.comment}</td>
-                                                <td style={{ width: "100px" }}>
-                                                    <button
-                                                        onClick={() =>
-                                                            handleViewScheduledTransaction(
-                                                                schedule.id,
-                                                            )
-                                                        }
-                                                        className="btn btn-primary btn-sm"
-                                                    >
-                                                        View
-                                                    </button>
-                                                </td>
-                                            </tr>
-                                        ))
-                                    ) : (
-                                        <tr>
-                                            <td
-                                                colSpan="5"
-                                                className="text-center text-muted"
-                                            >
-                                                No scheduled transactions
-                                                available
-                                            </td>
-                                        </tr>
-                                    )}
-                                </tbody>
-                            </table>
                         </div>
                     </div>
                 </div>
